+++ api-js/src/types/__tests__/dkim-results-connection.test.js
