+++ api-js/src/types/__tests__/dmarc.test.js
@@ -1,61 +1,74 @@
 const { DB_PASS: rootPass, DB_URL: url } = process.env
 
 const { ArangoTools, dbNameFromFile } = require('arango-tools')
+const { toGlobalId } = require('graphql-relay')
 const {
   GraphQLNonNull,
   GraphQLID,
-  GraphQLInt,
   GraphQLString,
+  GraphQLInt,
 } = require('graphql')
-const { toGlobalId } = require('graphql-relay')
-const { makeMigrations } = require('../../../../migrations')
-const { cleanseInput } = require('../../../validators')
-const {
-  domainLoaderByKey,
-  spfGuidanceTagConnectionsLoader,
-} = require('../../../loaders')
-const { spfType, domainType, guidanceTagConnection } = require('../index')
 
-describe('given the spfType object', () => {
-  describe('testing the field definitions', () => {
+const { makeMigrations } = require('../../../migrations')
+const { cleanseInput } = require('../../validators')
+const {
+  dmarcGuidanceTagConnectionsLoader,
+  domainLoaderByKey,
+} = require('../../loaders')
+const { dmarcType, domainType, guidanceTagConnection } = require('../index')
+
+describe('given the dmarcType object', () => {
+  describe('testing its field definitions', () => {
     it('has an id field', () => {
-      const demoType = spfType.getFields()
+      const demoType = dmarcType.getFields()
 
       expect(demoType).toHaveProperty('id')
       expect(demoType.id.type).toMatchObject(GraphQLNonNull(GraphQLID))
     })
     it('has a domain field', () => {
-      const demoType = spfType.getFields()
+      const demoType = dmarcType.getFields()
 
       expect(demoType).toHaveProperty('domain')
       expect(demoType.domain.type).toMatchObject(domainType)
     })
     it('has a timestamp field', () => {
-      const demoType = spfType.getFields()
+      const demoType = dmarcType.getFields()
 
       expect(demoType).toHaveProperty('timestamp')
       expect(demoType.timestamp.type).toMatchObject(GraphQLString)
     })
-    it('has a lookups field', () => {
-      const demoType = spfType.getFields()
+    it('has a dmarcPhase field', () => {
+      const demoType = dmarcType.getFields()
 
-      expect(demoType).toHaveProperty('lookups')
-      expect(demoType.lookups.type).toMatchObject(GraphQLInt)
+      expect(demoType).toHaveProperty('dmarcPhase')
+      expect(demoType.dmarcPhase.type).toMatchObject(GraphQLInt)
     })
     it('has a record field', () => {
-      const demoType = spfType.getFields()
+      const demoType = dmarcType.getFields()
 
       expect(demoType).toHaveProperty('record')
       expect(demoType.record.type).toMatchObject(GraphQLString)
     })
-    it('has a spfDefault field', () => {
-      const demoType = spfType.getFields()
+    it('has a pPolicy field', () => {
+      const demoType = dmarcType.getFields()
 
-      expect(demoType).toHaveProperty('spfDefault')
-      expect(demoType.spfDefault.type).toMatchObject(GraphQLString)
+      expect(demoType).toHaveProperty('pPolicy')
+      expect(demoType.pPolicy.type).toEqual(GraphQLString)
+    })
+    it('has a spPolicy field', () => {
+      const demoType = dmarcType.getFields()
+
+      expect(demoType).toHaveProperty('spPolicy')
+      expect(demoType.spPolicy.type).toMatchObject(GraphQLString)
+    })
+    it('has a pct field', () => {
+      const demoType = dmarcType.getFields()
+
+      expect(demoType).toHaveProperty('pct')
+      expect(demoType.pct.type).toMatchObject(GraphQLInt)
     })
     it('has a guidanceTags field', () => {
-      const demoType = spfType.getFields()
+      const demoType = dmarcType.getFields()
 
       expect(demoType).toHaveProperty('guidanceTags')
       expect(demoType.guidanceTags.type).toMatchObject(
@@ -64,8 +77,8 @@ describe('given the spfType object', () => {
     })
   })
 
-  describe('testing the field resolvers', () => {
-    let query, drop, truncate, migrate, collections, domain, spf, spfGT
+  describe('testing its field resolvers', () => {
+    let query, drop, truncate, migrate, collections, domain, dmarc, dmarcGT
 
     beforeAll(async () => {
       ;({ migrate } = await ArangoTools({ rootPass, url }))
@@ -79,20 +92,22 @@ describe('given the spfType object', () => {
         domain: 'test.domain.gc.ca',
         slug: 'test-domain-gc-ca',
       })
-      spf = await collections.spf.save({
+      dmarc = await collections.dmarc.save({
         timestamp: '2020-10-02T12:43:39Z',
-        lookups: 5,
+        dmarcPhase: 1,
         record: 'txtRecord',
-        spfDefault: 'default',
-        guidanceTags: ['spf1'],
+        pPolicy: 'pPolicy',
+        spPolicy: 'spPolicy',
+        pct: 100,
+        guidanceTags: ['dmarc1'],
       })
-      await collections.domainsSPF.save({
+      await collections.domainsDMARC.save({
         _from: domain._id,
-        _to: spf._id,
+        _to: dmarc._id,
       })
-      spfGT = await collections.spfGuidanceTags.save({
-        _key: 'spf1',
-        tagName: 'SPF-TAG',
+      dmarcGT = await collections.dmarcGuidanceTags.save({
+        _key: 'dmarc1',
+        tagName: 'DMARC-TAG',
         guidance: 'Some Interesting Guidance',
         refLinksGuide: [
           {
@@ -117,27 +132,18 @@ describe('given the spfType object', () => {
       await drop()
     })
 
-    describe('testing the id resolver', () => {
+    describe('testing the id field resolver', () => {
       it('returns the resolved value', () => {
-        const demoType = spfType.getFields()
+        const demoType = dmarcType.getFields()
 
-        expect(demoType.id.resolve({ id: '1' })).toEqual(toGlobalId('spf', '1'))
+        expect(demoType.id.resolve({ id: '1' })).toEqual(toGlobalId('dmarc', 1))
       })
     })
     describe('testing the domain resolver', () => {
       it('returns the resolved value', async () => {
-        const demoType = spfType.getFields()
+        const demoType = dmarcType.getFields()
 
         const loader = domainLoaderByKey(query, '1', {})
-
-        const expectedResult = {
-          _id: domain._id,
-          _key: domain._key,
-          _rev: domain._rev,
-          id: domain._key,
-          domain: 'test.domain.gc.ca',
-          slug: 'test-domain-gc-ca',
-        }
 
         await expect(
           demoType.domain.resolve(
@@ -145,66 +151,88 @@ describe('given the spfType object', () => {
             {},
             { loaders: { domainLoaderByKey: loader } },
           ),
-        ).resolves.toEqual(expectedResult)
+        ).resolves.toEqual({
+          _id: domain._id,
+          _key: domain._key,
+          _rev: domain._rev,
+          id: domain._key,
+          domain: 'test.domain.gc.ca',
+          slug: 'test-domain-gc-ca',
+        })
       })
     })
     describe('testing the timestamp resolver', () => {
       it('returns the resolved value', () => {
-        const demoType = spfType.getFields()
+        const demoType = dmarcType.getFields()
 
         expect(
           demoType.timestamp.resolve({ timestamp: '2020-10-02T12:43:39Z' }),
         ).toEqual('2020-10-02T12:43:39Z')
       })
     })
-    describe('testing the lookups resolver', () => {
+    describe('testing the dmarcPhase resolver', () => {
       it('returns the resolved value', () => {
-        const demoType = spfType.getFields()
+        const demoType = dmarcType.getFields()
 
-        expect(demoType.lookups.resolve({ lookups: 1 })).toEqual(1)
+        expect(demoType.dmarcPhase.resolve({ dmarcPhase: 1 })).toEqual(1)
       })
     })
     describe('testing the record resolver', () => {
       it('returns the resolved value', () => {
-        const demoType = spfType.getFields()
+        const demoType = dmarcType.getFields()
 
         expect(demoType.record.resolve({ record: 'txtRecord' })).toEqual(
           'txtRecord',
         )
       })
     })
-    describe('testing the spfDefault resolver', () => {
+    describe('testing the pPolicy resolver', () => {
       it('returns the resolved value', () => {
-        const demoType = spfType.getFields()
+        const demoType = dmarcType.getFields()
 
-        expect(
-          demoType.spfDefault.resolve({ spfDefault: 'spfDefault' }),
-        ).toEqual('spfDefault')
+        expect(demoType.pPolicy.resolve({ pPolicy: 'pPolicy' })).toEqual(
+          'pPolicy',
+        )
+      })
+    })
+    describe('testing the spPolicy resolver', () => {
+      it('returns the resolved value', () => {
+        const demoType = dmarcType.getFields()
+
+        expect(demoType.spPolicy.resolve({ spPolicy: 'spPolicy' })).toEqual(
+          'spPolicy',
+        )
+      })
+    })
+    describe('testing the pct resolver', () => {
+      it('returns the resolved value', () => {
+        const demoType = dmarcType.getFields()
+
+        expect(demoType.pct.resolve({ pct: 100 })).toEqual(100)
       })
     })
     describe('testing the guidanceTags resolver', () => {
       it('returns the resolved value', async () => {
-        const demoType = spfType.getFields()
+        const demoType = dmarcType.getFields()
 
-        const loader = spfGuidanceTagConnectionsLoader(
+        const loader = dmarcGuidanceTagConnectionsLoader(
           query,
           '1',
           cleanseInput,
           {},
         )
-
-        const guidanceTags = ['spf1']
+        const guidanceTags = ['dmarc1']
 
         const expectedResult = {
           edges: [
             {
-              cursor: toGlobalId('guidanceTags', spfGT._key),
+              cursor: toGlobalId('guidanceTags', dmarcGT._key),
               node: {
-                _id: spfGT._id,
-                _key: spfGT._key,
-                _rev: spfGT._rev,
-                id: spfGT._key,
+                _id: dmarcGT._id,
+                _key: dmarcGT._key,
+                _rev: dmarcGT._rev,
                 guidance: 'Some Interesting Guidance',
+                id: dmarcGT._key,
                 refLinksGuide: [
                   {
                     description: 'refLinksGuide Description',
@@ -217,8 +245,8 @@ describe('given the spfType object', () => {
                     ref_link: 'www.refLinksTechnical.ca',
                   },
                 ],
-                tagId: 'spf1',
-                tagName: 'SPF-TAG',
+                tagId: 'dmarc1',
+                tagName: 'DMARC-TAG',
               },
             },
           ],
@@ -226,8 +254,8 @@ describe('given the spfType object', () => {
           pageInfo: {
             hasNextPage: false,
             hasPreviousPage: false,
-            startCursor: toGlobalId('guidanceTags', spfGT._key),
-            endCursor: toGlobalId('guidanceTags', spfGT._key),
+            startCursor: toGlobalId('guidanceTags', dmarcGT._key),
+            endCursor: toGlobalId('guidanceTags', dmarcGT._key),
           },
         }
 
@@ -235,7 +263,7 @@ describe('given the spfType object', () => {
           demoType.guidanceTags.resolve(
             { guidanceTags },
             { first: 1 },
-            { loaders: { spfGuidanceTagConnectionsLoader: loader } },
+            { loaders: { dmarcGuidanceTagConnectionsLoader: loader } },
           ),
         ).resolves.toEqual(expectedResult)
       })
