+++ api-js/src/types/__tests__/domain-status.test.js
@@ -1,4 +1,4 @@
-const { StatusEnum } = require('../../../enums')
+const { StatusEnum } = require('../../enums')
 const { domainStatus } = require('../domain-status')
 
 describe('given the domainStatus object', () => {
