+++ api-js/src/types/__tests__/email-scan.test.js
@@ -2,14 +2,14 @@ const { DB_PASS: rootPass, DB_URL: url } = process.env
 
 const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { toGlobalId } = require('graphql-relay')
-const { makeMigrations } = require('../../../../migrations')
-const { cleanseInput } = require('../../../validators')
+const { makeMigrations } = require('../../../migrations')
+const { cleanseInput } = require('../../validators')
 const {
   domainLoaderByKey,
   dkimLoaderConnectionsByDomainId,
   dmarcLoaderConnectionsByDomainId,
   spfLoaderConnectionsByDomainId,
-} = require('../../../loaders')
+} = require('../../loaders')
 const {
   emailScanType,
   domainType,
