+++ api-js/src/types/dmarc-report/detail-tables/__tests__/detail-tables.test.js
