+++ api-js/src/types/dmarc-report/detail-tables/__tests__/spf-failure-table.test.js
